backend/requirements.txt
Flask==3.1.0
Flask-CORS==5.0.0
Flask-Session==0.8.0
pandas==2.2.3
python-dotenv==1.0.1
requests==2.32.3